scripts/requirements.txt
 
# scripts/requirements.txt
pandas
scikit-learn
imbalanced-learn
matplotlib
seaborn
xgboost
mlflow
prefect
boto3
joblib
python-dotenv